resources_old/requirements.txt
certifi
charset-normalizer
contourpy
cycler
filelock
fonttools
idna
Jinja2
joblib
kiwisolver
MarkupSafe
matplotlib
mpmath
networkx
numpy
packaging
Pillow
pyparsing
python-dateutil
PyYAML
requests
scikit-learn
scipy
six
sympy
threadpoolctl
torch
torchvision
typing_extensions
urllib3
pandas
torchinfo